Render the console with DrawLists
In order to figure out how many lines a console line takes up, you
essentially follow the same steps as if you were actually rendering
the line without actually rendering it.  This isn't necessarily taxing
on the CPU, but requires that you either duplicate code, or have a weird
function that renders or returns the height.

I chose instead to implement a DrawList.  A DrawList is a vector of
functions and their essential parameters (called edicts) that are meant
to be iterated through and rendered all in one batch.  Creating the
DrawList is much like rendering a bunch of related items in sequence,
except you add edicts to the list instead of actually calling the real
renderer functions.  At the time the DrawList is complete, you should
also be aware of the true combined width and height of the entire batch
and can thereafter set the width and height of the DrawList itself.

Rendering of the console is simple at this point.  Iterate backwards
through the console buffer and if a line doesn't have a DrawList,
generate one.  Then, check the height of the DrawList and if you can
fit it, render it.  Repeat until you hit the top of the screen.