Introduce Graphics
Graphics are resource handles that are handed out by the
GraphicsManager class.  In and of themselves, they contain no actual
picture data, just some metadata about its size and offset.  However,
they can be passed to the renderer, which will draw the appropriate
picture.

Also, I suppose I started the conversion of the status bar to using the
OpenGL renderer too, but I forgot to commit that work earlier, so...