Introduce world resolution setter function
Now, instead of being capable of rendering to an arbitrary width and
height inside the video buffer, we always render to the full width and
height of the buffer and resize the buffer instead.

I also discovered that my getter functions for my buffers were not
being properly inlined because they weren't in the header, so I moved
them out of the .cpp file.