Add horizontal FOV support
You can tweak the FOV variable in R_InitTextureMapping and get a wider
view now, without weird floor/ceiling textures and incorrectly scaled
sprites.  Thanks again to prBoom for helping me out.

I'm not so sure if making projection equal to the focal length is a
wise idea though, given that the focal length is inaccurate due to
inaccuracies in the finetangent table.  Also, don't push the FOV past
120 degrees unless you want to see weird unrendered glitches.